backend/requirements.txt
appdirs==1.4.4
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.17
django-allauth==0.43.0
django-cors-headers==3.5.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-mssql-backend==2.8.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
importlib-metadata==2.0.0
importlib-resources==3.3.0
oauthlib==3.1.0
Pillow==8.0.1
pipenv==2020.11.15
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.1.0
virtualenv-clone==0.5.4
zipp==3.4.0
